+++ src/honestRFTree.h
@@ -57,7 +57,7 @@ public:
 
   void selectBestFeature(
     size_t& bestSplitFeature,
-    float& bestSplitValue,
+    long double& bestSplitValue,
     float& bestSplitLoss,
     std::vector<size_t>* featureList,
     std::vector<size_t>* averagingSampleIndex,
